meta-intel: meta-isg: openssl-qat: fix for patch error
openssl-qat is different from other packages in the way that it
has a patch present in a zip file that is unpacked and applied
on top of openssl. This results in patch errors when doing an
incremental build using bitbake.

To resolve, splitted do_patch into two - one to unpack the qat patch
and apply it on openssl. And, other to apply the patches added in
SRC_URI.

(From meta-intel rev: 3fdd6fdae1df3223d4d748dbcf609c27a79b8f9d)

Signed-off-by: Rahul Kumar Gupta <rahul.kumarxx.gupta@intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>